+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/ContributionProvidersBoxTestGenerated.java
@@ -1868,12 +1868,6 @@ public class ContributionProvidersBoxTestGenerated extends AbstractContributionP
       }
 
       @Test
-      @TestMetadata("AssistedParamNamesDisabled.kt")
-      public void testAssistedParamNamesDisabled() {
-        runTest("compiler-tests/src/test/data/box/inject/assisted/AssistedParamNamesDisabled.kt");
-      }
-
-      @Test
       @TestMetadata("AssistedParametersInLaterOrderWork.kt")
       public void testAssistedParametersInLaterOrderWork() {
         runTest("compiler-tests/src/test/data/box/inject/assisted/AssistedParametersInLaterOrderWork.kt");