+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/AssistedFactoryTransformer.kt
@@ -456,8 +456,6 @@ internal class AssistedFactoryTransformer(
               param.toAssistedParameterKey(
                 symbols = context.metroSymbols,
                 typeKey = substitutedTypeKey,
-                useAssistedParamNamesAsIdentifiers =
-                  context.options.useAssistedParamNamesAsIdentifiers,
               )
             },
         )